requirements.txt
google-genai>=1.0.0
groq>=0.12.0
ollama>=0.4.0
PyQt6>=6.6.0
python-docx>=1.1.0
PyPDF2>=3.0.0
psutil>=5.9.0
duckduckgo-search>=7.0.0
